+++ carp_mobile_sensing/lib/probes/sensors/light_probe.dart
@@ -28,5 +28,6 @@ class LightProbe extends BufferingPeriodicStreamProbe {
     super.resume();
   }
 
+  @override
   void onData(luxValue) => luxValues.add(luxValue);
 }