+++ carp_mobile_sensing_app/lib/src/app.dart
@@ -1,0 +1,13 @@
+part of mobile_sensing_app;
+
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData.dark(),
+      home: Scaffold(
+        body: ProbesList(),
+      ),
+    );
+  }
+}