+++ console_app/android/app/src/main/java/dk/cachet/consoleapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.carpmobilesensingapp;
+package dk.cachet.consoleapp;
 
 import android.os.Bundle;
 import io.flutter.app.FlutterActivity;