+++ carp_mobile_sensing/CHANGELOG.md
@@ -2,6 +2,7 @@
 * major updates to the domain model as a `core` library
     * `Measure` now have a `configuration`
 * simplification to probe implementations
+* all probes now uses Streams and supports a reactive programming model